fix(MOD-72): create crates for all task need build check
