+++ Cargo.toml
@@ -2,15 +2,28 @@
 name = "mira-war-of-realms"
 version = "0.1.0"
 edition = "2024"
+authors = ["Daniel Ramke <daniel.ramke@outlook.com>"]
+exclude = ["dist", "build", "assets", "credits", "conf"]
 
 [[bin]]
 name = "mira-war-of-realms"
 path = "src/main.rs"
 
+[workspace]
+members = [
+    "crates/ui_lib",
+    "crates/audio_lib",
+    "crates/environment_lib",
+    "crates/entities_lib",
+    "crates/system"
+]
+
 [profile.release]
+panic = "unwind"
+opt-level = "s"
 lto = true
 codegen-units = 1
-panic = "unwind"
+strip = true
 
 [profile.dev]
 opt-level = 1
@@ -20,11 +33,6 @@ opt-level = 3
 debug = true
 
 [dependencies]
-# Audio
-bevy_kira_audio = { version = "^0.22.0", features = ["ogg"] }
-
-# Camera
-bevy_third_person_camera = {version = "0.2.0"}
 
 # Debug
 bevy-inspector-egui = {version = "0.29.1" }
@@ -37,17 +45,23 @@ fluent-bundle = {version = "^0.15"}
 # Phisyc
 bevy_rapier3d = {version = "0.28.0", features = ["simd-nightly", "debug-render-3d", "default"]}
 
-# Utils
-rand = { version = "0.9.0" }
-toml = {version = "^0.8.0", features = ["parse", "display"]}
-serde = { version = "1.0.217", features = ["derive"] }
-regex = { version = "^1.11" }
+# Internal Crates
+system = { path = "crates/system" }
+ui_lib = { path = "crates/ui_lib" }
+audio_lib = { path = "crates/audio_lib" }
+environment_lib = { path = "crates/environment_lib" }
+entities_lib = { path = "crates/entities_lib" }
 
 [dependencies.bevy]
 version = "0.15.2"
-default-features = true
+default-features = false
 features = [
-    "jpeg",
-    "png",
-    "dynamic_linking"
+    "bevy_asset",
+    "bevy_core_pipeline",
+    "bevy_render",
+    "bevy_window",
+    "bevy_winit",
+    "default_font",
+    "multi_threaded",
+    "x11"
 ]
