+++ crates/audio_lib/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "audio_lib"
+version = "0.0.1"
+edition = "2024"
+authors = ["Daniel Ramke <daniel.ramke@outlook.com>"]
+
+[lib]
+name = "audio_lib"
+path = "src/lib.rs"
+
+[dependencies]
+bevy_kira_audio = { version = "^0.22.0", features = ["ogg"] }
+
+system = { path = "../system" }
+
+[dependencies.bevy]
+version = "0.15.2"
+default-features = false
+features = [
+    "bevy_asset",
+    "bevy_core_pipeline",
+    "multi_threaded",
+]
