+++ crates/environment_lib/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "environment_lib"
+version = "0.0.1"
+edition = "2024"
+authors = ["Daniel Ramke <daniel.ramke@outlook.com>"]
+
+[lib]
+name = "environment_lib"
+path = "src/lib.rs"
+
+[dependencies]
+
+regex = { version = "^1.11" }
+bevy_rapier3d = {version = "0.28.0", features = ["simd-nightly", "debug-render-3d", "default"]}
+
+system = { path = "../system" }
+
+[dependencies.bevy]
+version = "0.15.2"
+default-features = false
+features = [
+    "animation",
+    "bevy_animation",
+    "bevy_asset",
+    "bevy_color",
+    "bevy_core_pipeline",
+    "bevy_gltf",
+    "bevy_pbr",
+    "bevy_render",
+    "bevy_scene",
+    "multi_threaded",
+    "png",
+    "jpeg",
+    "tonemapping_luts"
+]
