+++ crates/environment_lib/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod environment;
