+++ crates/system/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod states;
+pub mod utils;
+pub mod config;
+pub mod events;
+pub mod service;
+pub mod commons;
+
+pub const PLAYER_VOID_THRESHOLD: f32 = -100.0;
